net-psp: reintroduce psp encap in the protocol layer
Encapsulation in the netdevice is simpler than in the TCP stack.
Edge cases such as software segmentation do not require custom psp
handlers with this model.

For real devices, define this flag and encap in the driver
(analogous to how all drivers have to decap in the driver).

For packetdrill testing we need to maintain the previous in-stack
approach, however. Because it reads packets not using tun_put_user,
but using packet sockets, which intercept before device can encap.

Bring back in-stack encapsulation. Select between the two modes as
a compile time option CONFIG_INET_PSP_ENCAP_IN_DEV. In-dev encap
has not been tested with this patch. For testing in-dev encap, just
do not include this patch in your patch series.

Signed-off-by: Willem de Bruijn <willemb@google.com>